split artifact identity into core microcrate
