+++ crates/uselesskey-core-artifact-id/README.md
@@ -1,0 +1,11 @@
+# uselesskey-core-artifact-id
+
+Core artifact identity primitives shared across uselesskey microcrates.
+
+This microcrate is intentionally focused on modeling fixture identity:
+
+- `ArtifactDomain`
+- `DerivationVersion`
+- `ArtifactId`
+
+It does **not** perform seed parsing or derivation by itself.
